+++ src/com/steftmax/temol/render/Updatable.java
@@ -3,5 +3,5 @@ package com.steftmax.temol.render;
 public interface Updatable {
 	
 	
-	public void update(final long delta);
+	public void update(final float delta);
 }
